Extract CompetitionsCell test helpers to a separate file
